Try to fix matlab bindings...
